requirements.txt
django~=5.0.4
djangorestframework~=3.15.1
django-environ~=0.11.2
django-split-settings~=1.3.1
psycopg2-binary~=2.9.5